+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 　　1、增加php dcrphp list/help/-h等功能  
 　　2、增加查看某个命令的帮助如:php dcrphp help console:make  
 　　3、文件管理及在线代码编辑  
+　　4、增加一种短地址实现功能:config/route/web.php    
 
 1.0.5(20200807)  
 　　1、修正创建命令行没有返回值的bug